+++ src/App/Http/Controllers/OrderController.php
@@ -1,5 +1,6 @@
 <?php namespace Redooor\Redminportal\App\Http\Controllers;
 
+use Input;
 use Exception;
 use Validator;
 use Redooor\Redminportal\App\Http\Traits\SorterController;
@@ -104,77 +105,47 @@ class OrderController extends Controller
 
     public function postStore()
     {
-        $sid = \Input::get('id');
-
-        $rules = array(
-            'product_id'        => 'required_without_all:bundle_id,pricelist_id',
-            'bundle_id'         => 'required_without_all:product_id,pricelist_id',
-            'pricelist_id'      => 'required_without_all:product_id,bundle_id',
-            'transaction_id'    => 'required',
-            'payment_status'    => 'required',
-            'paid'              => 'numeric',
-            'email'             => 'required|email'
-        );
-
-        $validation = \Validator::make(\Input::all(), $rules);
-
-        $redirect_url = (isset($sid)) ? 'admin/orders/edit/' . $sid : 'admin/orders/create';
+        $redirect_url = 'admin/orders/create';
+        
+        $validation = $this->validateStoreInputs(Input::all());
         
         if ($validation->fails()) {
             return redirect($redirect_url)->withErrors($validation)->withInput();
         }
         
-        $transaction_id = \Input::get('transaction_id');
-        $payment_status = \Input::get('payment_status');
-        $paid           = \Input::get('paid');
-        $email          = \Input::get('email');
+        $transaction_id = Input::get('transaction_id');
+        $payment_status = Input::get('payment_status');
+        $paid           = Input::get('paid');
+        $email          = Input::get('email');
         
-        $apply_to_models = array();
-        
-        // Save products to order
-        $products = \Input::get('product_id');
-        if (count($products) > 0) {
-            foreach ($products as $item) {
-                $model = Product::find($item);
-                if ($model != null) {
-                    $apply_to_models[] = $model;
-                }
-            }
-        }
-        // Save bundles to order
-        $bundles = \Input::get('bundle_id');
-        if (count($bundles) > 0) {
-            foreach ($bundles as $item) {
-                $model = Bundle::find($item);
-                if ($model != null) {
-                    $apply_to_models[] = $model;
-                }
-            }
-        }
-        // Save pricelists to order
-        $pricelists = \Input::get('pricelist_id');
-        if (count($pricelists) > 0) {
-            foreach ($pricelists as $item) {
-                $model = Pricelist::find($item);
-                if ($model != null) {
-                    $apply_to_models[] = $model;
-                }
-            }
-        }
-
-        // No product/bundle to add
-        if (count($apply_to_models) == 0) {
-            $errors = new \Illuminate\Support\MessageBag;
-            $errors->add('productError', "The items selected may have been deleted. Please try again.");
-            return redirect($redirect_url)->withErrors($errors)->withInput();
-        }
-        
+        // Check user first
         $user = User::where('email', $email)->first();
         
         if ($user == null) {
             // No such user
             $errors = new \Illuminate\Support\MessageBag;
             $errors->add('userError', "The user may have been deleted. Please try again.");
+            return redirect($redirect_url)->withErrors($errors)->withInput();
+        }
+        
+        $apply_to_models = array();
+        
+        // Save products to order
+        $selected_products = Input::get('selected_products');
+        $apply_to_models = array_merge($apply_to_models, $this->addModelToArray($selected_products, new Product));
+        
+        // Save bundles to order
+        $selected_bundles = Input::get('selected_bundles');
+        $apply_to_models = array_merge($apply_to_models, $this->addModelToArray($selected_bundles, new Bundle));
+        
+        // Save pricelists to order
+        $pricelists = Input::get('pricelist_id');
+        $apply_to_models = array_merge($apply_to_models, $this->addModelToArraySimpleMode($pricelists, new Pricelist));
+        
+        // No product/bundle to add
+        if (count($apply_to_models) == 0) {
+            $errors = new \Illuminate\Support\MessageBag;
+            $errors->add('productError', "The items selected may have been deleted. Please try again.");
             return redirect($redirect_url)->withErrors($errors)->withInput();
         }
         
@@ -191,9 +162,80 @@ class OrderController extends Controller
         }
         
         // Save coupons to order
-        $coupons = \Input::get('coupon_id');
+        $coupons = Input::get('coupon_id');
+        $errors = $this->addCouponToOrder($coupons, $new_order);
+        
+        if ($errors) {
+            return redirect($this->pageRoute)->withErrors($errors);
+        }
+        
+        return redirect($this->pageRoute);
+    }
+    
+    /**
+     * Add model to array, content is JSON string
+     *
+     * @param array $models
+     * @param object $model Model
+     *
+     * @return array
+     **/
+    private function addModelToArray($models, $model)
+    {
+        $apply_to_models = array();
+        
+        if (count($models) > 0) {
+            foreach ($models as $item_json) {
+                $item = json_decode($item_json);
+                $object = $model->find($item->id);
+                if ($object != null) {
+                    for ($count = 1; $count <= $item->quantity; $count++) {
+                        $apply_to_models[] = $object;
+                    }
+                }
+            }
+        }
+        
+        return $apply_to_models;
+    }
+    
+    /**
+     * Add model to array, simple mode, content is NOT JSON string
+     *
+     * @param array $models
+     * @param object $model Model
+     *
+     * @return array
+     **/
+    private function addModelToArraySimpleMode($models, $model)
+    {
+        $apply_to_models = array();
+        
+        if (count($models) > 0) {
+            foreach ($models as $item) {
+                $object = $model->find($item);
+                if ($object != null) {
+                    $apply_to_models[] = $object;
+                }
+            }
+        }
+        
+        return $apply_to_models;
+    }
+    
+    /**
+     * Add coupons to order
+     *
+     * @param array $coupons
+     * @param object $new_order Order
+     *
+     * @return array
+     **/
+    private function addCouponToOrder($coupons, $new_order)
+    {
+        $errors = new \Illuminate\Support\MessageBag;
+        
         if (count($coupons) > 0) {
-            $errors = new \Illuminate\Support\MessageBag;
             foreach ($coupons as $item) {
                 $model = Coupon::find($item);
                 if ($model != null) {
@@ -211,13 +253,40 @@ class OrderController extends Controller
             
             // Set coupon discount
             $new_order->setDiscounts();
-            
-            if (count($errors) > 0) {
-                return redirect('admin/orders')->withErrors($errors);
-            }
         }
         
-        return redirect('admin/orders');
+        return $errors;
+    }
+    
+    /**
+     * Validates inputs before storing data
+     *
+     * @param array $inputs
+     *
+     * @return object Validator
+     **/
+    private function validateStoreInputs($inputs)
+    {
+        $rules = array(
+            'selected_products' => 'required_without_all:selected_bundles,pricelist_id',
+            'selected_bundles'  => 'required_without_all:selected_products,pricelist_id',
+            'pricelist_id'      => 'required_without_all:selected_products,selected_bundles',
+            'transaction_id'    => 'required',
+            'payment_status'    => 'required',
+            'paid'              => 'numeric',
+            'email'             => 'required|email'
+        );
+        
+        $messages = [
+            'selected_products.required_without_all' =>
+                trans('redminportal::messages.order_error_selected_products_required_without_all'),
+            'selected_bundles.required_without_all' =>
+                trans('redminportal::messages.order_error_selected_bundles_required_without_all'),
+            'pricelist_id.required_without_all' =>
+                trans('redminportal::messages.order_error_pricelist_id_required_without_all')
+        ];
+
+        return Validator::make($inputs, $rules, $messages);
     }
     
     public function getUpdate($field = null, $sid = null, $status = null)