+++ CHANGELOG.md
@@ -69,6 +69,7 @@ This update introduces some new tables and lots of changes.
 20. Added reusable Revision and Revisionable trait for tracking of changes in any model. (issue #138)
 21. Added UI for viewing Revision History for Orders.
 22. Allow User to change their password and names (issue #159)
+23. Create new Order with product quantities (issue #165)
 
 ### Enhancements:
 1. Category model missing relationships with other models. (issue #140)
@@ -110,6 +111,7 @@ This update introduces some new tables and lots of changes.
 18. Product variant should inherit category from main product (issue #160)
 19. Move printMenu to Html class and partial template (issue #155)
 20. Add field automatically_apply to Coupon model (issue #162)
+21. UI: Orders Items list should group into same product and show quantity (issue #166)
 
 ### Bug fixes:
 1. User deletion: prevent user from deleting or deactivating own account while they are logged in. (issue #136)